backend/services/core_api/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-environ==0.11.2
psycopg[binary]==3.2.3
gunicorn==23.0.0
cryptography==44.0.0
boto3==1.35.78
requests==2.32.3
